[CWS_WIDI_UX] Set WiFi Display P2P go_intent to 14.
In order to be able to improve Performencies,
we need to be able to choose the P2P channel
used for WiFi Display connection. Only the
P2P Group Owner can choose it, so force the
go_intent to 14.

Change-Id: Ibe3266fcfada04c51ce0d5514f7be3c2fc835090
Category: aosp improvement
Domain: CWS.WIFI-Perf
Origin: internal
Upstream-Candidate: no, not-aosp
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3723
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: https://android.intel.com:443/283635