HEADER: fix the GTT mapping for 64-bits
jira: 6710

For 64 bit userspace, the GTT mapping structures should be able to store
64 bit handles in gtt_key. Moreover user virtual addresses can be 64
bits, and therefore the virtual address field in the syscall should be
wide enough to accept it.

Therefore, structures psb_gtt_mapping_arg and intel_dc_plane_ctx are
amended. Regression on 32 bit userspace might happen, and no tests have
been done so far to check it.

Change-Id: Ia621a7ca6a61f60ecb2d63ea61eb226aa88b26fc
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6710
Signed-off-by: Robert Jarzmik <robert.jarzmik@intel.com>
Signed-off-by: Tan Wei <weix.z.tan@intel.com>
Reviewed-on: https://android.intel.com:443/294277