Overlaps aosp TCP buffersize for hspap & lte bearer technology
To get maximum tPut it is critical to use optimal TCP send and
receive socket size for the bearer link you are using. The buffer
size shall be computed based on the following formula:
buffer size = bandwidth * RTT, the bandwidth is equivalent to the
maximum physical tPut for the established bearer. RTT round trip
time will be determined thanks to ping. for HSPAP bearer, the
default settings degrades DL by 1Mbits/s (cat26). for LTE with
cat6 and carrier aggregation, the default settings are under
sized based on lab measurement. see IMINAN-1282 for measurement
results.

Change-Id: Ie94abd6e71715f05715152e50a4929231a3f0466
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1282
Signed-off-by: Pascal Bedouet <pascal.bedouet@intel.com>
Reviewed-on: https://android.intel.com:443/285892