[AUDIO] Supports of HOTWORD input source
Since L-dessert, AudioHAL may be requested to open an input stream
with the hotword source, breaking the simple conversion to a mask.

It has been decided to assign the 16th bit to hotword.
This patch updates the definition of the input sources criterion
to take this new input source into account.

Change-Id: Icc9d43f7f297cbb653f3826eba3a930d13bfc25d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-6999
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/294055