SELinux policies for ASF.
A new domain called "asf" is created. This domain maps to ASF
securityfileservice. This domain is created as an "unconfined"
domain, as the securityfileservice needs to access every file
in the file system. If the asf domain were to be created as
a regular confined domain, then we would have a long list of
rules to allow it access every possible file in the file system,
and we might not be able to uncover all rules needed until after
product ships since not all possible file types are present in
a fresh flashed device.

Change-Id: I2a3365b4063cc545854718bd7dc62742b97c16dd
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-8267
Signed-off-by: Zhong Chen <zhong.chen@intel.com>
Reviewed-on: https://android.intel.com:443/304127